taglib: Remove compilation warnings

+++ projet-musique/taglib/toolkit/tutils.h
@@ -206,7 +206,7 @@ namespace TagLib
         // The last resort. May cause a buffer overflow.
 
         length = vsprintf(buf, format, args);
-        if(length >= BufferSize) {
+        if(static_cast<size_t>(length) >= BufferSize) {
           debug("Utils::formatString() - Buffer overflow! Returning an empty string.");
           length = -1;
         }
